submitting: add examples of referencing equations and figures
I suggest \autoref here because I think it's the easiest to use and
create a consistent convention going forward.  One can use \eqref or
raw \ref, but these lead to greater variation in individual styles.

I have also dropped ``Gala`` in favor of `Gala` because the behavior
is equivalent and simpler seems better.