Jed/submitting with refs (#692)
* submitting: add examples of referencing equations and figures

I suggest \autoref here because I think it's the easiest to use and
create a consistent convention going forward.  One can use \eqref or
raw \ref, but these lead to greater variation in individual styles.

I have also dropped ``Gala`` in favor of `Gala` because the behavior
is equivalent and simpler seems better.

* submitting: add example of fenced code block with syntax highlighting

I don't know how to protect the fence residing inside a larger fenced
region, thus there is a trailing TAB character that suppresses Sphinx
from interpreting as the end of the outer region.