Submitting.md: add a TiKZ example
Until now, readers were not informed about their access to the powerful
TiKZ library _inside_ their own paper. This might lead them to using
rasterized graphics inside their publication which cannot compete with the
infinite resolution of vector graphics.

Also, some might try hacking different methods to produce the illustrations
and inject them inside the 'paper.md' file with the markdown syntax. This
leads to inconsistent typesetting inside the final paper.

With this commit, I've put a TiKZ example taken from the pgfmanual
documentation for a demonstration. This example enjoys the same typesetting
as the rest of the paper. For example, the figure caption has the same font
as the body.